refactor(ui): extract <Transcript> pure renderer
