+++ ui/index.html
@@ -121,6 +121,7 @@
 <script type="text/babel" src="components/channels.jsx"></script>
 <script type="text/babel" src="components/channel_rules.jsx"></script>
 <script type="text/babel" src="components/chat/use-transcript.js"></script>
+<script type="text/babel" src="components/chat/transcript.jsx"></script>
 <script type="text/babel" src="components/chat/conversation.jsx"></script>
 <script type="text/babel" src="components/chats.jsx"></script>
 <script type="text/babel" src="components/harness_form.jsx"></script>
